@@ -10,10 +10,10 @@
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.1.3",
     "firebase": "^9.6.10",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-bootstrap": "^2.2.3",
     "react-bootstrap-carousel": "^4.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-hook-form": "^7.27.1",
     "react-icons": "^4.3.1",
     "react-redux": "^7.2.6",
