@@ -1,5 +1,5 @@
-import ReactDOM from 'react-dom';
 import React from "react";
+import { createRoot } from 'react-dom/client';
 
 import './global.scss';
 import App from './App';
@@ -7,13 +7,15 @@ import {BrowserRouter} from "react-router-dom";
 import {Provider} from "react-redux";
 import store from "./store";
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
     <BrowserRouter>
         <Provider store={store}>
             <React.StrictMode>
                 <App/>
             </React.StrictMode>
         </Provider>
-    </BrowserRouter>,
-    document.getElementById('root')
+    </BrowserRouter>
 );
