@@ -2,6 +2,7 @@ import classes from './Checkout.module.scss';
 import {useForm} from "react-hook-form";
 import {useSelector} from "react-redux";
 import {useEffect, useState} from "react";
+import Input from "../../../component/input/Input";
 
 const Checkout = () => {
     const user = useSelector(store => store.login.user);
@@ -23,6 +24,18 @@ const Checkout = () => {
 
     const {reset, register, handleSubmit, formState: {errors}} = useForm();
     // debugger
+    const items = [{
+        name: 'email',
+        value: tempUser?.email,
+        patternValue: /\S+@\S+\.\S+/,
+        message: 'Entered value does not match email format.'
+    }, {
+        name: 'password',
+        placeholder: 'Password',
+        type: 'password',
+        minLengthValue: 6,
+        minLengthMessage: 'Min length is 6',
+    }]
 
     const onSubmit = (data) => {
         debugger
@@ -33,29 +46,24 @@ const Checkout = () => {
             <div className={classes['billing-detail']}>
                 <h4 className={`text-uppercase`}>billing details</h4>
                 {tempUser && <form onSubmit={handleSubmit(onSubmit)}>
-                    <div>
-                        <input
-                            placeholder='You name here...'
-                            {...register('name', {
-                                required: 'Field Required',
-                                value: tempUser.name
-                            })}
+                    {items.map(item =>
+                        <Input
+                            key={item.name}
+                            errors={errors}
+                            register={register}
+                            required={true}
+                            value={item.value}
+                            width={100}
+                            placeholder={item.placeholder}
+                            label={item.name}
+                            type={item.type}
+                            ccsClass={item.disabled ? 'disabled' : ''}
+                            patternValue={item.patternValue}
+                            patternErrorMessage={item.message}
+                            minLengthValue={item.minLengthValue}
+                            minLengthMessage={item.minLengthMessage}
                         />
-                    </div>
-                    <div>
-                        <input
-                            placeholder='You email here...'
-                            {...register('email', {
-                                required: 'Field Required',
-                                value: tempUser.email,
-                                pattern: {
-                                    value: /\S+@\S+\.\S+/,
-                                    message: 'Entered value does not match email format.'
-                                }
-                            })}
-                        />
-                        {errors.email && <span role='alert'>{errors.email.message}</span>}
-                    </div>
+                    )}
 
                     <button type='submit'>click</button>
                 </form>}
